Get tints of brand colors

get_brand_color_tint() generates a range of tints (color variations) for a specific brand color, from black (position 0) through the brand color (position 500) to white (position 1000), and returns the hexadecimal color code at the specified position.

Usage

get_brand_color_tint(position = 500, color = "primary")

Arguments

position

(Optional) A numeric vector indicating the position of the brand color in the range of tints. The range of positions is from 0 to 1000 (Default: 500).

color

A character string indicating the name of a color present in the color section of the _brand.yml file (Default: "primary").

Value

A character vector with hexadecimal color codes.

Details

Path to _brand.yml

brandr will always look for a _brand.yml file in the root directory of your project. If the file is not found, an error message will be displayed. You can also set the path to the file manually using the options() function:

options(BRANDR_BRAND_YML = [PATH TO _brand.yml])

Brand Color Scales

To control the colors for each brand color scale, assign the desired hexadecimal color codes in a character vector to the following options:

BRANDR_COLOR_SEQUENTIAL: For sequential color scales

BRANDR_COLOR_DIVERGING: For diverging color scales

BRANDR_COLOR_QUALITATIVE: For qualitative color scales

You can use get_brand_color() to get the hexadecimal color codes from the _brand.yml file.

Example:

options(
  BRANDR_COLOR_SEQUENTIAL =
    get_brand_color(c("primary", "secondary")),
  BRANDR_COLOR_DIVERGING =
    get_brand_color(c("primary", "white", "secondary")),
  BRANDR_COLOR_QUALITATIVE =
    get_brand_color(c("primary", "secondary", "tertiary"))
)

See also

Examples

seq(0, 1000, 250)
#> [1]    0  250  500  750 1000
#> [1]    0  250  500  750 1000 # Expected

get_brand_color_tint(seq(0, 1000, 250), color = "primary")
#> [1] "#000000" "#6D271E" "#DA4E3C" "#ECA69D" "#FFFFFF"
#> [1] "#000000" "#6D271E" "#DA4E3C" "#ECA69D" "#FFFFFF" # Expected