Brand color scales for `ggplot2“

scale_*_brand_*() functions provide color scales for ggplot2 based on brand colors defined in the _brand.yml file. These functions create discrete, continuous, or binned scales with sequential, diverging, or qualitative color palettes that match your brand identity.

Usage

scale_brand(
  aesthetics = "color",
  scale_type = "c",
  color_type = "seq",
  direction = 1,
  na.value = NA,
  reverse = FALSE,
  brand_yml = here::here("_brand.yml"),
  ...
)

scale_color_brand_d(
  aesthetics = "color",
  scale_type = "d",
  color_type = "seq",
  direction = 1,
  ...
)

scale_color_brand_c(
  aesthetics = "color",
  scale_type = "c",
  color_type = "seq",
  direction = 1,
  ...
)

scale_color_brand_b(
  aesthetics = "color",
  scale_type = "b",
  color_type = "seq",
  direction = 1,
  ...
)

scale_colour_brand_d(
  aesthetics = "color",
  scale_type = "d",
  color_type = "seq",
  direction = 1,
  ...
)

scale_colour_brand_c(
  aesthetics = "color",
  scale_type = "c",
  color_type = "seq",
  direction = 1,
  ...
)

scale_colour_brand_b(
  aesthetics = "color",
  scale_type = "b",
  color_type = "seq",
  direction = 1,
  ...
)

scale_fill_brand_d(
  aesthetics = "fill",
  scale_type = "d",
  color_type = "seq",
  direction = 1,
  ...
)

scale_fill_brand_c(
  aesthetics = "fill",
  scale_type = "c",
  color_type = "seq",
  direction = 1,
  ...
)

scale_fill_brand_b(
  aesthetics = "fill",
  scale_type = "b",
  color_type = "seq",
  direction = 1,
  ...
)

Arguments

aesthetics

(Optional) A character string indicating the name of the aesthetic of the scale (e.g., "color"``, "fill"`).

scale_type

(Optional) A character string indicating the type of scale: "d"/"discrete", "c"/"continuous", or "b"/"binned".

color_type

(Optional) A character string indicating the type of color scale: "seq"/"sequential"``, "div"/"diverging"“, or "qual"/`"qualitative"` (Default: `c `).

direction

(Optional) Sets the order of colors in the scale. If 1 colors are as output by the palette function. If -1, the order of colors is reversed (Default: 1).

na.value

(Optional) A character string indicating the color to use for missing values. It must contain a hexadecimal color code or one of the values in colors() (Default: NA).

reverse

A logical indicating whether the legend or color bar should be reversed (Default: FALSE).

brand_yml

(Optional) A character string indicating the path to the _brand.yml file. If not provided, the function will look for a _brand.yml file in the current working directory.

...

Additional arguments passed to the ggplot2 scale function: discrete_scale(), continuous_scale(), or binned_scale().

Value

A ggplot2 scale object.

Details

To control the colors for each brand color scale, assign the desired hexadecimal color codes in a character vector to the following options:

BRANDR_COLOR_SEQUENTIAL for sequential color scales

BRANDR_COLOR_DIVERGING for diverging color scales

BRANDR_COLOR_QUALITATIVE for qualitative color scales

Use get_brand_color() to get the hexadecimal color codes from the _brand.yml file.

Example:

options(
  BRANDR_COLOR_SEQUENTIAL =
    get_brand_color(c("primary", "secondary")),
  BRANDR_COLOR_DIVERGING =
    get_brand_color(c("primary", "white", "secondary")),
  BRANDR_COLOR_QUALITATIVE =
    get_brand_color(c("primary", "secondary", "tertiary"))
)

Examples

if (base::interactive()) {
  library(ggplot2)

  ggplot(mtcars, aes(mpg, wt, color = cyl)) +
    geom_bar() +
    scale_color_brand_c() +
    theme_bw()

  ggplot(mtcars, aes(mpg, wt, fill = factor(cyl))) +
    geom_boxplot() +
    scale_fill_brand_d(color_type = "qual") +
    theme_bw()
}